Joint life tables (on first death) for now have a maximum age taken from the first insured person (i.e. the first table).
The reason is a purely technical one: The ages of the insured are not part of the table, so the getOmega function does not know any ages. Only the deathProbabilities function gets the actual ages.

This makes it impossible to create tables, where e.g. the last value is repeated as needed. Instead, we will always cut off the table at the maximum age of the table for the first insured person.